+++ CHANGELOG.md
@@ -17,6 +17,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * Now `initFromStorage` returns `CanPromise` `super/i-block/modules/state`
 * Removed the default export of submodules `core/dom`
+* Removed `StaticRouteMeta.entryPoint` and `StaticRouteMeta.dynamicDependencies` `core/router`
 * `build`:
   * Renamed `entries.webpack` to `graph.webpack`
   * Renamed `build.webpack` to `helpers.webpack`
@@ -29,6 +30,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
   * Added `entry.webpack`
 
 * Added API to load the dynamic dependencies `iBlock`
+* Added `StaticRouteMeta.load` `core/router`
 
 #### :house: Internal